refactor(datasouce/docker): split file
